refactor(locations->stops): Rename all location to stop
This commit renames all variables that have the string
`location` in the name. This was done to avoid confunsion
between what is a stop and what is a location.

Location in this context is something that has coordinates , an id and a
display name. Stop, is some location that the user wants to visit in a
route, including the starting point and the end point of said route.
This commit also implements a way to visualize existing routes on the
map